LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-scroll-functional-notation.tentative-expected.txt


FAIL animation-timeline: scroll(nearest) assert_equals: expected "50px" but got "50.18px"
FAIL animation-timeline: scroll(root) assert_equals: expected "50px" but got "50.169998px"
FAIL animation-timeline: scroll(self) assert_equals: expected "100px" but got "50.169998px"
FAIL animation-timeline: scroll(self), on non-scroller assert_equals: expected "none" but got "50.220001px"
FAIL animation-timeline: scroll(inline) assert_equals: expected "100px" but got "50.169998px"
FAIL animation-timeline: scroll(x) assert_equals: expected "100px" but got "50.279999px"
FAIL animation-timeline: scroll(y) assert_equals: expected "100px" but got "50.169998px"
